+++ test/unit/core/test-assert.js
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import {USER_ERROR_SENTINEL} from '../../../src/core/error-message-helpers';
+import {USER_ERROR_SENTINEL} from '../../../src/core/error/message-helpers';
 import {devAssert, userAssert} from '../../../src/core/assert';
 
 describes.sandboxed('assertions', {}, () => {
